requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
google-auth==2.32.0
requests==2.32.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
razorpay==1.4.2
sqlalchemy[asyncio]==2.0.32
asyncpg==0.29.0
email-validator==2.2.0
setuptools>=69.0.0
runpod>=1.6.0

# ML/AI Dependencies (torch installed in Dockerfile from official index)
# torch==2.2.1  <- Installed via Dockerfile from PyTorch index
# torchvision==0.17.1  <- Installed via Dockerfile from PyTorch index
numpy<2.0,>=1.26.0
diffusers>=0.31.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.2
peft>=0.12.0
pillow==10.1.0
huggingface-hub>=0.25.0
